LayoutTests/imported/w3c/web-platform-tests/uievents/click/click_event_target_child_parent-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: unimplemented
Click targeting when targets of down and up are child and parents

This test verifies that click event always goes to the first common ancestor of down and up event targets.
Press down the primary button on link 1 and move onto green container and release.
Press down the primary button on red container and move onto link 2 and release.
Click done.
link1
link2
Done

Harness Error (FAIL), message = Unhandled rejection: unimplemented

NOTRUN Click targets the nearest common ancestor
